+++ ethereum/spec/src/testFixtures/java/tech/pegasys/teku/spec/propertytest/suppliers/state/BuilderSupplier.java
@@ -11,17 +11,17 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.spec.datastructures.epbs.versions.gloas;
+package tech.pegasys.teku.spec.propertytest.suppliers.state;
 
 import tech.pegasys.teku.spec.SpecMilestone;
-import tech.pegasys.teku.spec.datastructures.state.versions.gloas.BuilderPendingWithdrawal;
+import tech.pegasys.teku.spec.datastructures.state.versions.gloas.Builder;
+import tech.pegasys.teku.spec.datastructures.state.versions.gloas.BuilderPendingPayment;
 import tech.pegasys.teku.spec.propertytest.suppliers.DataStructureUtilSupplier;
 import tech.pegasys.teku.spec.util.DataStructureUtil;
 
-public class BuilderPendingWithdrawalSupplier
-    extends DataStructureUtilSupplier<BuilderPendingWithdrawal> {
+public class BuilderSupplier extends DataStructureUtilSupplier<Builder> {
 
-  public BuilderPendingWithdrawalSupplier() {
-    super(DataStructureUtil::randomBuilderPendingWithdrawal, SpecMilestone.GLOAS);
+  public BuilderSupplier() {
+    super(DataStructureUtil::randomBuilder, SpecMilestone.GLOAS);
   }
 }
